+++ Settings.cs
@@ -1,4 +1,7 @@
 ﻿using BepInEx.Configuration;
+using static UnityEngine.ParticleSystem.PlaybackState;
+using static UnityEngine.UIElements.UIR.Allocator2D;
+using System;
 
 namespace BetterSpectator
 {
@@ -16,7 +19,7 @@ namespace BetterSpectator
             {
                 isClockEnabled = config.Bind("General", "Show Clock", true, "Shows the time of day when spectating");
                 isCauseOfDeathEnabled = config.Bind("General", "Show Cause Of Death", true, "Shows each players cause of death when spectating");
-                isChatEnabled = config.Bind("General", "Show Chat", true, "Shows the chat box and allows dead players to talk to each other");
+                isChatEnabled = config.Bind("General", "Show Chat", true, "Enable text chat when spectating, but only with other dead players(and see alive players text chat if you are spectating them, or if they are in range of the spectated player/using a walkie talkie)");
                 isExtraInputEnabled = config.Bind("General", "Enable Extra Input", true, "Enables using the scroll wheel/arrow keys to scroll forwards/back between alive players when spectating");
                 isLoaded = true;
             }