+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v1.2:
-- Still be able to text chat but only with other dead players (and see alive players text chat if you are spectating them, or they are in range of the spectated player)
-- All options configurable via LethalConfig (or any other config mod)
+- Enable text chat while spectating, but only with other dead players (and see alive players text chat if you are spectating them, or they are in range of the spectated player/using a walkie talkie)
+- Options to enable/disable different features have been added, and are now configurable via LethalConfig (or any other config mod/or by manually changing the config files)
 
 ## v1.1:
 - Added ability to scroll forward and back through specated players when dead (using scroll wheel and arrow keys)